pyproject.toml
[project]
name = "atmega128-interface"
version = "0.1.0"
description = "Student-friendly Python interface for ATmega128 microcontroller"
authors = [
    {name = "SOC3050 Course", email = "course@university.edu"}
]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "pyserial>=3.5",
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

web = [
    "plotly>=5.15.0",
    "dash>=2.12.0",
    "gunicorn>=21.0.0"
]

iot = [
    "paho-mqtt>=1.6.0",
    "boto3>=1.28.0",
    "azure-iot-device>=2.12.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"